# api/openapi.json
{
  "components": {
    "schemas": {
      "AIConsentRequest": {
        "properties": {
          "version": {
            "maximum": 10000.0,
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version"
        ],
        "title": "AIConsentRequest",
        "type": "object"
      },
      "AIConsentResponse": {
        "properties": {
          "action": {
            "enum": [
              "granted",
              "withdrawn"
            ],
            "title": "Action",
            "type": "string"
          },
          "recorded_at": {
            "format": "date-time",
            "title": "Recorded At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version",
          "action",
          "recorded_at"
        ],
        "title": "AIConsentResponse",
        "type": "object"
      },
...
      "AdmissionErrorDetail": {
        "properties": {
          "code": {
            "enum": [
              "generation_disabled",
              "generation_admission_unavailable",
              "generation_rate_limited",
              "guest_admission_unavailable",
              "guest_session_rate_limited",
              "stream_admission_unavailable",
              "stream_limit_reached"
            ],
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "AdmissionErrorDetail",
        "type": "object"
      },
      "AdmissionErrorResponse": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/AdmissionErrorDetail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "AdmissionErrorResponse",
        "type": "object"
      },
      "AppleIdentityRequest": {
        "properties": {
          "identity_token": {
            "maxLength": 16000,
            "minLength": 100,
            "title": "Identity Token",
            "type": "string"
          }
        },
        "required": [
          "identity_token"
        ],
        "title": "AppleIdentityRequest",
        "type": "object"
      },
      "AuthTokenResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "refresh_expires_in": {
            "title": "Refresh Expires In",
            "type": "integer"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "token_type": {
            "const": "bearer",
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "access_token",
          "refresh_token",
          "expires_in",
          "refresh_expires_in"
        ],
        "title": "AuthTokenResponse",
        "type": "object"
      },
...
      "DeleteMyDataRequest": {
        "properties": {
          "confirmation": {
            "const": "DELETE",
            "title": "Confirmation",
            "type": "string"
          }
        },
        "required": [
          "confirmation"
        ],
        "title": "DeleteMyDataRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RefreshTokenRequest": {
        "properties": {
          "refresh_token": {
            "maxLength": 512,
            "minLength": 32,
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshTokenRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Itinera API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/ai-consent": {
      "delete": {
        "operationId": "withdraw_ai_consent_api_v1_auth_ai_consent_delete",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIConsentResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Withdraw Ai Consent",
        "tags": [
          "auth"
        ]
      },
      "post": {
        "operationId": "grant_ai_consent_api_v1_auth_ai_consent_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIConsentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIConsentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Grant Ai Consent",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/apple": {
      "post": {
        "operationId": "sign_in_with_apple_api_v1_auth_apple_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppleIdentityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Invalid Apple identity token"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "Apple sign-in is unavailable"
          }
        },
        "summary": "Sign In With Apple",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/apple/link": {
      "post": {
        "operationId": "link_apple_identity_api_v1_auth_apple_link_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppleIdentityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication or Apple identity token is invalid"
          },
          "409": {
            "description": "Apple identity belongs to another Itinera library"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "Apple sign-in is unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Link Apple Identity",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/guest": {
      "post": {
        "operationId": "create_guest_api_v1_auth_guest_post",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdmissionErrorResponse"
                }
              }
            },
            "description": "Guest-session admission limit reached",
            "headers": {
              "Retry-After": {
                "description": "Whole seconds until the client should retry.",
                "schema": {
                  "minimum": 1.0,
                  "type": "integer"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdmissionErrorResponse"
                }
              }
            },
            "description": "Guest-session admission cannot be evaluated",
            "headers": {
              "Retry-After": {
                "description": "Whole seconds until the client should retry.",
                "schema": {
                  "minimum": 1.0,
                  "type": "integer"
                }
              }
            }
          }
        },
        "summary": "Create Guest",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "delete": {
        "operationId": "delete_my_data_api_v1_auth_me_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteMyDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Account data is absent after deletion or safe replay"
          },
          "401": {
            "description": "Bearer token is invalid, or is expired while its account still exists"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete My Data",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "refresh_guest_session_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Guest Session",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}